requirements.txt
Django==1.11.13
djangorestframework==3.8.2
mysqlclient==1.3.1